メソッドの中身を雑にヘルパー関数として切り出した
