+++ src/js/game/dom-floaters/post-battle/procedures/hide.ts
@@ -1,0 +1,13 @@
+import { PostBattleFloaterProps } from "../props";
+
+/**
+ * フローターを非表示にする
+ * @param props プロパティ
+ */
+export function hide(props: PostBattleFloaterProps): void {
+  props.root.style.display = "none";
+  props.unsubscribers.forEach((v) => {
+    v.unsubscribe();
+  });
+  props.root.innerHTML = "";
+}
