+++ src/js/game/dom-floaters/post-battle/props.ts
@@ -1,4 +1,4 @@
-import { Subject } from "rxjs";
+import { Subject, Unsubscribable } from "rxjs";
 
 import { Exclusive } from "../../../exclusive/exclusive";
 import { PostBattle } from "../../post-battle";
@@ -11,4 +11,6 @@ export type PostBattleFloaterProps = {
   readonly exclusive: Exclusive;
   /** 選択完了通知 */
   readonly selectionComplete: Subject<PostBattle>;
+  /** アンサブスクライバ */
+  unsubscribers: Unsubscribable[];
 };
